--- a/SUPABASE_MIGRATION_GUIDE.md
+++ b/SUPABASE_MIGRATION_GUIDE.md
@@ -1,0 +1,55 @@
+# 🚀 Supabase Migration Guide for Sully Booking System
+
+This guide documents the complete migration from Prisma to Supabase for the Sully booking system.
+
+## 📋 Migration Overview
+
+**From**: Prisma ORM with PostgreSQL
+**To**: Supabase (PostgreSQL with built-in auth, real-time, and APIs)
+
+### Key Benefits
+- ✅ Built-in authentication and authorization
+- ✅ Real-time subscriptions
+- ✅ Automatic API generation
+- ✅ Row Level Security (RLS)
+- ✅ Better performance for serverless deployments
+- ✅ Reduced bundle size
+- ✅ Built-in file storage
+
+## 🗄️ Database Schema
+
+The migration preserves all existing functionality while optimizing for Supabase:
+
+### Tables Created
+1. **users** - User accounts with roles and subscription info
+2. **venues** - Restaurant/venue information
+3. **bookings** - Reservation data
+4. **venue_opening_hours** - Operating hours for each venue
+5. **venue_availability** - Daily capacity and availability
+
+### Security Features
+- Row Level Security (RLS) enabled on all tables
+- Role-based access control
+- Secure API endpoints
+- JWT-based authentication
+
+## ✅ Migration Checklist
+
+- [x] Database schema migrated
+- [x] Supabase client configured
+- [x] Authentication system updated
+- [x] API routes converted
+- [x] RLS policies implemented
+- [x] Environment variables configured
+- [x] Testing completed
+- [x] Documentation updated
+- [x] Deployment configuration updated
+
+---
+
+**Migration Status**: ✅ Complete
+**Performance**: 🚀 Improved
+**Security**: 🔐 Enhanced
+**Maintainability**: 📈 Better
+
+The Sully booking system has been successfully migrated to Supabase with improved performance, security, and maintainability!