--- a/app/.env.example
+++ b/app/.env.example
@@ -1,44 +1,28 @@
+# Supabase Configuration
+NEXT_PUBLIC_SUPABASE_URL=your_supabase_project_url
+NEXT_PUBLIC_SUPABASE_ANON_KEY=your_supabase_anon_key
+SUPABASE_SERVICE_ROLE_KEY=your_supabase_service_role_key
 
-# Database Configuration
-# For production, use external database like PlanetScale, Supabase, or Railway
-DATABASE_URL="postgresql://username:password@host:port/database"
-
-# Authentication
-JWT_SECRET="your-secure-jwt-secret-key-here"
+# Application Configuration
+NEXT_PUBLIC_APP_URL=http://localhost:3000
+NODE_ENV=development
 
 # Stripe Configuration
-# Get these from your Stripe Dashboard (https://dashboard.stripe.com/apikeys)
-STRIPE_SECRET_KEY="sk_live_..." # or sk_test_... for testing
-STRIPE_PUBLISHABLE_KEY="pk_live_..." # or pk_test_... for testing
-STRIPE_WEBHOOK_SECRET="whsec_..." # From Stripe webhook configuration
-
-# Stripe Price IDs
-# Create these in your Stripe Dashboard under Products
-STRIPE_PAID_PRICE_ID="price_..." # Your paid plan price ID
-STRIPE_PREMIUM_PRICE_ID="price_..." # Your premium plan price ID
-
-# Environment Configuration
-NODE_ENV="production" # Use "development" for local development
-STRIPE_MODE="live" # Use "test" for development/testing
-
-# App Configuration
-NEXT_PUBLIC_APP_URL="https://your-app-name.netlify.app" # Your Netlify app URL
-WEBHOOK_BASE_URL="https://your-app-name.netlify.app" # Same as above for webhooks
-
-# Optional: Disable Next.js telemetry
-NEXT_TELEMETRY_DISABLED=1
-
-# File Upload Configuration (for Netlify)
-# You may want to use external file storage like Cloudinary or AWS S3
-NEXT_PUBLIC_UPLOAD_PROVIDER="netlify" # or "cloudinary", "aws"
+STRIPE_SECRET_KEY=sk_test_...
+NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=pk_test_...
+STRIPE_WEBHOOK_SECRET=whsec_...
 
 # Email Configuration (Optional)
-# Configure if you want to send emails
-EMAIL_FROM="noreply@yourdomain.com"
-EMAIL_PROVIDER="smtp" # or "sendgrid", "mailgun"
+SMTP_HOST=smtp.gmail.com
+SMTP_PORT=587
+SMTP_USER=your-email@gmail.com
+SMTP_PASS=your-app-password
 
-# SMTP Configuration (if using SMTP)
-SMTP_HOST="smtp.your-provider.com"
-SMTP_PORT="587"
-SMTP_USER="your-smtp-username"
-SMTP_PASS="your-smtp-password"
+# Admin Configuration
+ADMIN_EMAIL=admin@sully-booking.com
+ADMIN_PASSWORD=admin123
+
+# Security
+JWT_SECRET=your-jwt-secret-key
+NEXTAUTH_SECRET=your-nextauth-secret
+NEXTAUTH_URL=http://localhost:3000